I made homework to point 9 , 10 and 11 is not ready yet

+++ 11_Generics_And_Utility_Classes/homework11/src/main/kotlin/Ammo.kt
@@ -1,17 +1,17 @@
 enum class Ammo(
    name: String,
-  val damage: Double,
+  val damage: Int,
    val criticalChanceDamage: Int,
     val criticalDamageRatio: Double
 ) {
-    pistolCartrige("pistol_ammo",20.0, 12, 1.2),
-    sniperCartrige("sniper_ammo",85.0, 90, 1.4),
-    assaultCartrige("assault_ammo",70.0, 40, 1.5);
+    pistolCartrige("pistol_ammo",20, 12, 1.2),
+    sniperCartrige("sniper_ammo",85, 90, 1.4),
+    assaultCartrige("assault_ammo",70, 40, 1.5);
 
 
-    fun takingCurrentDamage(): Double {
+    fun takingCurrentDamage(): Int {
         return if (criticalChanceDamage.probability())
-            damage * criticalDamageRatio
+            damage * criticalDamageRatio.toInt()
         else damage
     }
